# Proj1/Rest/openAPI specification.json
{
  "definitions": {},
  "info": {
    "description": "powered by Flasgger",
    "termsOfService": "/tos",
    "title": "A swagger API",
    "version": "0.0.1"
  },
  "paths": {
    "/iot/{ts}/{device}": {
      "delete": {
        "parameters": [
          {
            "description": "Timestamp",
            "in": "path",
            "name": "ts",
            "required": true,
            "type": "string"
          },
          {
            "description": "Device ID",
            "in": "path",
            "name": "device",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "204": {
            "description": "Successfully deleted the IoT reading"
          },
          "404": {
            "description": "IoT reading not found"
          }
        },
        "summary": "Delete a specific IoT reading by ID"
      },
      "get": {
        "parameters": [
          {
            "description": "Timestamp",
            "in": "path",
            "name": "ts",
            "required": true,
            "type": "string"
          },
          {
            "description": "Device ID",
            "in": "path",
            "name": "device",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Return the specified IoT reading"
          },
          "404": {
            "description": "IoT reading not found"
          }
        },
        "summary": "Get a specific IoT reading by ID"
      },
      "put": {
        "parameters": [
          {
            "description": "Timestamp",
            "in": "path",
            "name": "ts",
            "required": true,
            "type": "string"
          },
          {
            "description": "Device ID",
            "in": "path",
            "name": "device",
            "required": true,
            "type": "string"
          },
          {
            "example": {
              "co": 0.00495593864839124,
              "humidity": 51.0,
              "light": true,
              "lpg": 0.00765082227055719,
              "motion": true,
              "smoke": 0.0204112701224129,
              "temp": 25.0
            },
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "properties": {
                "co": {
                  "type": "number"
                },
                "humidity": {
                  "type": "number"
                },
                "light": {
                  "type": "boolean"
                },
                "lpg": {
                  "type": "number"
                },
                "motion": {
                  "type": "boolean"
                },
                "smoke": {
                  "type": "number"
                },
                "temp": {
                  "type": "number"
                }
              },
              "type": "object"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully updated the IoT reading"
          },
          "400": {
            "description": "Bad Request - Invalid input data"
          },
          "404": {
            "description": "IoT reading not found"
          }
        },
        "summary": "Update a specific IoT reading by ID"
      }
    },
    "/iots": {
      "get": {
        "responses": {
          "200": {
            "description": "Return all IoT readings"
          }
        },
        "summary": "Get all IoT readings"
      },
      "post": {
        "parameters": [
          {
            "example": {
              "co": 0.00495593864839124,
              "device": "b8:27:eb:bf:9d:51",
              "humidity": 51.0,
              "light": true,
              "lpg": 0.00765082227055719,
              "motion": true,
              "smoke": 0.0204112701224129,
              "temp": 25.0,
              "ts": "1.5945120943859746E9"
            },
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "properties": {
                "co": {
                  "type": "number"
                },
                "device": {
                  "type": "string"
                },
                "humidity": {
                  "type": "number"
                },
                "light": {
                  "type": "boolean"
                },
                "lpg": {
                  "type": "number"
                },
                "motion": {
                  "type": "boolean"
                },
                "smoke": {
                  "type": "number"
                },
                "temp": {
                  "type": "number"
                },
                "ts": {
                  "type": "string"
                }
              },
              "type": "object"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successfully created a new IoT reading"
          },
          "400": {
            "description": "Bad Request - Invalid input data"
          }
        },
        "summary": "Create a new IoT reading"
      }
    }
  },
  "swagger": "2.0"
}